01_indexing_qdrant/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.26
qdrant-client>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
